Rename proxy module to http proxy
